usm: http2: tests: migrated tests to usm packages
